requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
